+++ postfix/src/util/dict_regexp.c
@@ -599,7 +599,7 @@ static DICT_REGEXP_RULE *dict_regexp_parseline(const char *mapname, int lineno,
 	if (prescan_context.max_sub > first_exp->re_nsub) {
 	    msg_warn("regexp map %s, line %d: out of range replacement index \"%d\": "
 		     "skipping this rule", mapname, lineno,
-		     prescan_context.max_sub);
+		     (int) prescan_context.max_sub);
 	    FREE_EXPR_AND_RETURN(first_exp, 0);
 	}
 	if (second_pat.regexp != 0) {
